Subject:  Implemented Model Features
Recipe and Nutrition Model classes were implemented.  A RecipeManager class was developed to hold the logic for the Recipe model.  RecipeClient is an adjacent class to RecipeManager that actually goes out to the API to make network calls.  (RecipeManager class could disappear if a design decision is made to implement a facade for the ViewController to interact with to retrieve data).