Subject:  Implemented Controller Level Functionality and Some View Functionality
Expanded capability of RecipeAPI, which is a facade for the controller to interact with to obtain data and to perform operations on data as necessary.  Implemented view controller functions.  Implemented searcher with uiactivityindicatorview.  Also implemented a custom uitableviewcell.  These classes are currently skeletons and need to be upgraded for functionality.